+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 All notable changes to the "avalonia-project-helper" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
-## [Unreleased]
+## [0.0.1] - 25/02/2025
 
 - Initial release
+- Create Project command
+- Install Templates command